bme680-collector/app/requirements.txt
bme680>=2.0.0
paho-mqtt>=2.0.0
